Fix bug in test suite causing PS and PDF to fail with rgb24 and device offsets.
Patch from cworth@cworth.org.

The trick with this one is that the test suite is using device offsets
in a rather fabricated way, and the final copy from an offset surface
to a non-offset surface is between two surfaces of the same size, so
must be 1:1. To get this we turn the device offset off before the
copy.

In a more conventional use of device offsets the surfaces would be of
different sizes, and the offset would be desired during the copy.