SVG: Fix mask and try to support CAIRO_CONTENT_ALPHA.
Cairo and SVG differ in their implementation of mask. Cairo only uses alpha channel where SVG uses all channels. So, before using a surface for masking, we use a filter that sets RGB channels to 1.0.

Support for CAIRO_CONTENT_ALPHA is similar and use the same filter when a SVG alpha surface is composited.