Use fallbacks in PDF backend for CAIRO_ANTIALIAS_NONE.
Perhaps there's a way to preserve that hint in PDF output, but until
we have code that actually does that, the correct thing to do is to
call it unsupported and let the fallbacks do their thing.

With this commit, the two regressions that were recently introduced
now pass again. Specifically:

	rectangle-rounding-error
	unantialiased-shapes