[FreeType] Unset and set to FC_RGBA_NONE the FC_RGBA attribute on pattern
if we don't want it.  This stuff is tricky, but I hope to explain:  In your
fontconfig configuration, you may match on "font", or on "pattern".  Turning
subpixel on typically looks like:

        <match target="font">
		<test qual="all" name="rgba">
			<const>unknown</const>
		</test>
		<edit name="rgba" mode="assign"><const>rgb</const></edit>
	</match>

This works good enough, and if you set to ANTIALIAS_GRAY, this will not
override that.  Now one may forget to match on "font" target, or intentionally
match on the pattern.  That happens before cairo font options are substituted
in the pattern.  So, to give a hint of subpixel in your config file, you can
write:

        <match target="pattern">
		<edit name="rgba" mode="assign"><const>rgb</const></edit>
	</match>

You don't really need to check for current values, as FcConfigSubstitute is
run before merging cairo_font_options_t in.  What this patch does, is to reset
pattern's rgba property if the font options explicitly ask for ANTIALIAS_GRAY.
This is the only place in cairo-ft-font.c that we use FcPatternDel, so I
thought some explanation is needed.