Add new tessellator (unused) in cairo-bentley-ottmann.c
This is the implementation as it cooked in the new-tessellator branch
available from:

	git://people.freedesktop.org/~cworth/cairo

The file here comes from commit eee4faf79900be2c5fda1fddd49737681a9e37d6 in
that branch. It's sitting here not hooked up to anything in cairo yet,
and still with a main function with test cases, etc.