perf/paint: Make all sizes use a consistent number of iterations
This effectively reverts the change made in:

	3c407aa80a43e1724e732482ada5ad6718fe33ae

The justification is that we want to be able to see the
trends of increasing the image size while keeping the test
the same. But changing the number of iterations in a size-
specific way interferes with that. Also, the standard deviation
is almost always better this way, (though it now varies more
by size).

And as a bonus, the total time required to run the suite is now
less.