Fix repeating source surface patterns with xlib backend.
This broke with the clone_similar optimization in
8d7a02ed58e06584eb09575e6ca11d0a81094ab6 The optimization added an
interest rectangle to clone_similar, but with a repeating source
pattern, the interest rectangle might not intersect the extents of the
surface at all.

The test suite caught this with the trap-clip case.

The fix here is to clone the entire surface if the pattern has an
extend mode of REPEAT.