[PS] Eliminate compiler warnings about unoptimizable loops
An innocient-looking loop like this:

  for (j = 0; j <= last; j++)
    something();

cannot be optimized, because it may loop forever!
Imagine the case that last is MAXINT, the loop will never end.  The correct
way to write it is:

  for (j = 0; j < last+1; j++)
    something();

In this case, if last is MAXINT, the loop will never run.  Not correct, but
better than looping forever.

Still better would be to correctly handle the MAXINT case (even though it
doesn't make any sense to show MAXINT number of glyphs in one operation!)  To
do that, we can use the fact that the input num_glyphs is a signed.  If
there is one good thing about using signed int as input length, it's that you
can use an unsigned looping variable to avoid looping forever.  That is
exactly what this patch does.