[configure] Remove the pkg-config minimum version requirement
It was added in commit 157663e1fdf120a71c6cc8222b88d2915a5a73eb because there
was a serious bug in the pkg.m4 file from pkg-config 0.18.  However, it was
fixed two days later.  So, instead of requireing 0.18.1 or greater which is
not available on Nokia 770 SDK, we just ignore the broken pkg-config 0.18.
The problem only happens if one runs autogen.sh with against a broken
pkg-config.  Configuring tarballs is not affected.