[cairo-scaled-font] Skip invisible glyphs (like space) in glyph_extents() (#9422)
When computing extents for an array of glyphs, just taking min/max of x/y for
the bounding box of each glyph doesn't work.  The reason being that an
invisible glyph (like the space glyph) should not modify the resulting
extents, but it will.  So now we skip invisible glyphs.