Fix _cairo_pattern_acquire_surface_for_surface to not clone outside surface extents
Rotation and other transformations would cause extents to be
computed which were outside the bounds of the surface to be
cloned, (and for non repeating patterns). Now we simply
restrict the computed extents to the surface extents.

This fixes the xlib failure of the recent rotate-image-surface-paint
test, (the apparently distinct ps failure remains).