[boilerplate/test-paginated] Issue a show_page before writing buffer
This is needed since cairo-test does not issue a show_page now, and
the one coming from paginated layer is too late (around finish).

The way we force the show_page is kinda funny.  That is, we create
a temporary cairo_t and call cairo_show_page() on it.  But works.