[cairo-skiplist] Reduce MAX_LEVEL from 31 to 15
The probability that a node of level L is generated is
0.25^(L-1) * 0.75.  It means, a node of level 15 or
more will be used with a probability of about 3 * 10^-9.
That's really rare...

Actually that's not still true, because the level of a new
node is capped by current max-level plus one.  So to really
get a node with a level of 15 one should first get a node
of level 2, then 3, then 4, ..., finally 15.  Now that's
REALLY rare.

And guess what, the skiplist only start behaving bad with a
max level cap of MAX_LEVEL when having on the order of
4**MAX_LEVEL items in it.  I really hope we don't get there.