Change _cairo_dtostr to have a void return type
As previously implemented, there's no essential information in the
return value from _cairo_dotostr, (the caller can simply use strlen
to recompute the same value, which is what the only caller is already
doing).

There would be real information in a return value which would return
the result from the call to snprintf for the case where the buffer is
not large enough for the number being printed.