[xlib] The buggy_repeat path should be removed for Xsun now
Sun never released a version of Solaris with Xorg with the buggy repeat
problem.  This patch was only needed for development versions of Solaris
Nevada (roughly builds 25-30).  The latest S10U release and Nevada releases
have a fixed Xorg.  So no users should ever encounter this bug.

We need to remove this test because Xsun has the same VendorString and a lower
VendorRelease number so it falsely triggers buggy_repeat to be turned on.
They just added Xrender support recently to Xsun, so this wasn't an issue
before recently.