[cairo-xlib-surface]: flush work queue before sending glyphs to X server.
  After introducing a work queue for deferred destruction of X resource
my firefox crashes over and over again because XRenderFreeGlyphs is trying
to free a non-exist glyph (already freed). The problematic call sequence is
something like below:
  XRenderAddGlyphs (20990204, 20069)
  XRenderAddGlyphs (20990204, 20069)
  XRenderFreeGlyphs (20990204, 20069)
  XRenderFreeGlyphs (20990204, 20069)
You can see the two add/free glyphs is interlaced. And obviously, we'll crash
at the last one. To fix this bug, we must be ensure here's no pending work
to free the glyph that we want to sent.