[cairo-output-stream] Write out floats to 18 digits after decimal point
We write floats using %f as the scientific format used by smarter %g is
invalid in PS/PDF.  %f however by default rounds to five digits after
decimal point.  This was causing precision loss and making the newly
added degenerate-pen test fail for PDF.  We now print up to 18 digits
which is as many bits doubles can accomodate.  We can be smarter, but
that's for another commit.