Round floats in output-stream to 6 significant digits after decimal
The previous commit increased the precision of floats from 6 digits
after the decimal point to 18 digits to correct rounding errors with
very small numbers. However most of the time this extra precision is
not required and results in increased PS/PDF output.

This commit makes the precision after the decimal point 6 significant
digits. For example:

   1.234567
   0.123456
   0.00123456
   0.00000000123456