[quartz] zero memory on bitmap creation
7 quartz/argb32/similar tests were showing failures due to 
random bits being set. The problem turned out to be that the 
initial test surface was not cleared before the similar surface 
was created by push_group. The problem behaviour is more obvious
if you run the tests with MallocScribble=1.

Fix this by calling memset after bitmap allocation; the multiplication
here cannot overflow since we already checked that in malloc_ab.