PS: Make Type 3 fonts text selectable
Put the glyph names in the Encoding array the same as is done for
Type 1 and Type42 fonts.

Acroread and Evince are still unable to correctly extract the text
after conversion with ps2pdf. However examining the pdf file shows
the glyph names are correct so this is probably a limitation of
Acroread and Evince.