PDF: Emit all patterns after content stream
To fix this performance issue

http://lists.cairographics.org/archives/cairo/2007-December/012197.html

the PDF surface needs to avoid starting and stopping the content
stream every time it emits a pattern. This patch makes the PDF surface
store a list of all patterns used while the content stream is written
out then write out all the patterns after the content stream is
closed.