[cairo-font-options] Check for a NULL cairo_font_options_t
On IRC Drakou reported a user error whereby cairo_scaled_font_create()
was called with a NULL cairo_font_options_t. However, instead of
reporting the error back to the user, cairo instead segfaulted trying
to dereference the NULL pointer!

Add a guard to check that the options is not NULL.