PS: Fix to never break the final ~> that terminates a base85 stream.
This was recently broken in commit 40d5082c24819968a5ee5a8f72e3b9cd61cb6105
which uses a base85-stream but without strings for encapsulating image
data. The bug was that the protection for ~> was only being applied
to the string encodings rather than all base85 streams.

We got lucky recently that the 24.8 change just happened to put the ~>
sequence on the line-wrap boundary so the test suite tripped up the
bug. Otherwise, we could have missed this for quite some time.