Add _cairo_pdf_operators_flush()
The optimizations planned for pdf-operators will mean that it will no
longer emit complete operations on each call to
fill/stroke/show_glyphs. For example a call to _show_glyphs() may not
finish the text operation to allow a subsequent call to _show_glyphs()
to be merged into the same text object.

A flush function is required to force pdf_operators to complete the
current operation before the pdf surface can emit any pdf operators.