[cairo-pdf-operators] Fix backward cluster iteration
It's trickier than it appears...

There still remains the question of whether the loop should iterate
utf8 forward and glyphs backward (as it does now), or iterate glyphs
forward and utf8 backward.

Doing utf8 forward has the benefit that glyphs appear in the PDF stream
in the logical order of text.  That may be a good thing.

Doing glyphs forward has the benefit of most glyphs taking their natural
width, providing for a more compact PDF output.

Firefox 3 currently calls cairo_show_glyphs() with glyphs going
right-to-left for RTL runs.  That's like going utf8 forward.

On the other hand, if we are fine doing glyphs backward, then we may as
well remove the "backward" flag and ask callers of show_text_glyphs to
produce glyph array going right to left for RTL runs.  The idea behind
having a backward flag in the first place was to allow keeping utf8 in
logical order and glyphs in visual left-to-right order.  If we don't
do that in practice, there's no point in having the flag.  Another design
idea of the flag was to mark right-to-left runs explicitly so the backend
can use this information.  In that case we probably better rename it to
"rtl" as the current semantics make it perfectly fine to set a LTR segment
with backward flag on: just revers the glyph array...

Anyway, the way to determine the right thing to do is to see how Acrobat
extracts text for RTL text in logical vs visual order.  Then we can decide
what to do in the PDF backend and whether the flag is worth keeping.
Going to experiment with that.