[doc/public/check-doc-coverage.sh] Make doc building needed for the test faster
We used to do a full doc build to just to run the coverage test.  That's
way too slow to expect people to run regularly.  Instead now we just do the
source code scanning part of the doc build system that is just enough to
know if all symbols are documented.

A full doc build can be done as always by invoking "make doc", and indeed
will be called as part of "make dist" or "make distcheck".