[test] Add an exercise for masking an image using the ctm.
Quote Kai-Uwe Behrmann:
"The expected behaviour for masking in Cairo is to set the mask according
to the current active matrix and apply unchanged to a to be masked
surface.

In SVG the mask element is bound to the masked object and thus the local
matrix from that image object applies to the nested mask as well."

This is a test case to exercise applying a mask to an image under
separate transformations.

Original patch by Kai-Uwe Behrmann, altered to run the test against all
backends (where it causes poppler to crash on my machine <evil grin>).