[test/filter-bilinear-extents] Remove XFAIL.
With the reintroduction of Carl Worth's and Owen Taylor's work to expand
pattern extents as necessary to include extra samples due to the filter
radius, we expect filter-bilinear-extents to PASS. This patch series is
important as it factors out our assumptions about filter radius into a
single function and cleans up the code in its wake.

However, since the external PS/PDF renderers do not necessarily use the same
filter as cairo (and currently they only use NEAREST due to lack of
/Interpolate emission in those backends) we can expect differences in test
output. So add the respective reference images to capture current
expectations and to highlight future changes.

Fixes bugs:
  Bug 15349 - bad clipping with EXTEND_NONE
  [https://bugs.freedesktop.org/show_bug.cgi?id=15349],
  Bug 15367 -Improve filtering handling in cairo-pattern.c
  [https://bugs.freedesktop.org/show_bug.cgi?id=15367]