[configure.ac] Require autoconf >= 2.59 and automake >= 1.9.6
These are the versions available on RHEL5 (two years old now), and
we know cairo works with them.  There's evidence that our build system
does not work with older automake, and we've been requiring autoconf 2.58
but no one ever tested 2.58 with the new build system.  It's very likely
that 2.58 doesn't work and needs some macro backporting.  In any case,
no one reported that they have 2.58 when I asked on the list.