[tessellator] Replace open-coding _cairo_int64_cmp().
We often use the construct:
  if (_cairo_int64_lt (A, B)
     return -1;
  if (_cairo_int64_gt (A, B)
     return 1;
  return 0;
to compare two large integers (int64, or int128) which does twice the
required work on CPUs without large integer support. So replace it with a
single wideint function _cairo_int64_cmp() and therefore allow
opportunities to both shrink the code size and write a more efficient
comparison. (The primarily motivation is to simply replace each block with
a single more expressive line.)