[gl] Use spans directly as geometry instead of rasterizing to a temporary.
I was hoping for a bigger win than this, but cairogears is now significantly
faster than it was with just traps.  One potential problem is that adding
src texture coordinates blew up vertex size from 12 bytes to 20, and we're
emitting quite a few vertices.  There's plenty of hope for the future,
though: With shaders we could remove the 2 new texcoords again by stuffing a
position to source coordinate transformation matrix in constants.  Then
with EXT_geometry_shader4 plus EXT_gpu_shader4 we should be able to get the
24 bytes for 2 vertices down to 16 or less.