[cairo] Track the MRU scaled font
When observing applications two patterns emerge. The first is due to
Pango, which wraps each glyph run within a context save/restore. This
causes the scaled font to be evicted after every run and reloaded on the
next. This is caught by the MRU slot on the cairo_scaled_font_map and
prevents a relatively costly traversal of the hash table and holdovers.

The second pattern is by applications that directly manage the rendering
of their own glyphs. The prime example of this is gnome-terminal/vte. Here
the application frequently alternates between a few scaled fonts - which
requires a hash table retrieval every time.

By introducing a MRU slot on the gstate we are able to directly recover
the scaled font around 90% of the time.

Of 110,000 set-scaled-fonts:
  4,000 were setting the current font
 96,000 were setting to the previous font
  2,500 were recovered from the MRU on the cairo_scaled_font_map
  7,500 needed a hash retrieval
which compares to ~106,000 hash lookups without the additional MRU slot on
the gstate.

This translates to an elapsed time saving of ~5% when replaying a
gnome-terminal trace using the drm backend.