quartz: Don't fallback to pixman for repeating radial gradients.
Figuring out where the outer circle should move to is tricky. I hope the
algebra in there is understandable.

This is a nice performance improvement, probably because we avoid painting the
gradient over the entire clipBox (which is usually the entire surface).

I tried to write reftests that compared a repeating radial gradient to a
non-repeating gradient with manually repeated stops, but it didn't work because
the rasterization was slightly different --- I'm not sure why.

This patch also forces us to use pixman for all degenerate cases where the
circles intersect. This at least makes us consistent across platforms.

From https://bugzilla.mozilla.org/show_bug.cgi?id=508227