XCB: Fix for all unbounded operators
_cairo_xcb_surface_fixup_unbounded_boxes() calculated a list of boxes that it
has to clear to make an unbounded operator work correctly. Then it cleared the
boxes that were drawn instead of clearing the list of boxes that it has to
clear.

The reason that this wasn't noticed before is that there is an optimization in
case we have only one box instead of a whole list of boxes. This hid the bug.

This fixes the "unbounded-operator" test case.

Signed-off-by: Uli Schlachter <psychon@znc.in>