xcb: Remove a wrong optimization
Clear surfaces with an alpha channel are already replaced with a
transparent color in gstate, so this code was never hit.

Clear surfaces without an alpha channel, can be replaced with solid
black only if the surface pattern has an extend other than
CAIRO_EXTEND_NONE.

Fixes the clear-source test for xcb.

Signed-off-by: Uli Schlachter <psychon@znc.in>
Reviewed-by: Andrea Canciani <ranma42@gmail.com>