ps,pdf: Deal with empty domain gradients.
If all the stops of the gradient have the same offset and the
pattern's extend mode is EXTEND_PAD, then we cannot use the stops'
domain as the interpolation parameter range because this would produce
a gradient with the same start and end objects.  Such ranges tickle
bad behaviour in rasterisers.

We replace the color function with an appropriate step function
defined on [0 1].

Fixes radial-gradient-one-stop for pdf and ps3.

Reviewed-by: M Joonas Pihlaja <jpihlaja@cc.helsinki.fi>