arc-looping-dash: Add a missing cairo_restore ()
This test called cairo_save () without restoring again. Normally, this doesn't
cause a problem. However, when cairo-test-suite is called with "-a", it also
executes the tests on a similar surface. This then caused
cairo_pop_group_to_source to be called on a cairo_save state which causes a
CAIRO_STATUS_INVALID_RESTORE since commit 5d95ae924ed1520.

Signed-off-by: Uli Schlachter <psychon@znc.in>