tor: Inline reverse insertion sort for handling intersections
The majority of intersections are with the nearest neighbour only, or
within a few neighbours (in a dense intersection of lines) so if walk
the active list backwards and find the new place to insert upon an
intersection it is faster than performing a mergesort afterwards.

Given enough intersections, the win is quite huge (15-20% on many-strokes).

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>