configure.ac: remove annoying change of INSTALL file
Every time we run ./autogen.sh, the INSTALL file would be replace by
/usr/share/automake-1.11/INSTALL, which changes the file. This is very
annoying since it messes up the result of 'git diff' and git always
wants to commit it.

Declare it as foreign would fix this annoying issue. Since it's not a
GNU project, it makes sense to do so.

Signed-off-by: Yuanhan Liu <yuanhan.liu@linux.intel.com>